--- IoT/Oblig2/src/main/java/Display.java
@@ -1,7 +1,0 @@
-public class Display {
-
-    public void write(String message) {
-        System.out.println("DISPLAY: " + message);
-    }
-
-}
