+++ IoT/Oblig2/src/main/java/devices/Heating.java
@@ -1,3 +1,6 @@
+package devices;
+
+import controllers.Room;
 import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
 import org.eclipse.paho.client.mqttv3.MqttCallback;
 import org.eclipse.paho.client.mqttv3.MqttClient;
@@ -6,9 +9,15 @@ import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
 import org.eclipse.paho.client.mqttv3.persist.MemoryPersistence;
 
-public class Heating implements MqttCallback, Runnable {
+public class Heating implements MqttCallback {
 
     private Room room;
+
+    public static void main(String[] args) throws MqttException {
+
+            new Heating(new Room(20));
+
+    }
 
     public Heating (Room room)throws MqttException  {
         this.room = room;
@@ -44,11 +53,6 @@ public class Heating implements MqttCallback, Runnable {
 
     public void write (boolean newvalue) {
         room.actuate(newvalue);
-    }
-
-    @Override
-    public void run() {
-
     }
 
     /**
