+++ IoT/Oblig2/src/main/java/publishers/MQTTPubTemperature.java
@@ -1,3 +1,6 @@
+package publishers;
+
+import devices.TemperatureSensor;
 import org.eclipse.paho.client.mqttv3.MqttClient;
 import org.eclipse.paho.client.mqttv3.MqttConnectOptions;
 import org.eclipse.paho.client.mqttv3.MqttException;
